Add install_generator for configuration initializer and restructure files for readability.
