+++ lib/project_honeypot/railtie.rb
@@ -1,0 +1,10 @@
+module ProjectHoneypot
+  class Railtie < Rails::Railtie
+    initializer "project_honeypot.configure_rails_initialization" do
+      Rails.application.middleware.use ProjectHoneypot::Middleware
+    end
+    generators do
+      require "generators/install_generator"
+    end
+  end
+end
